requirements.txt
aws-cdk-lib==2.110.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.110.1a0
